Add files via upload
In this project, we want to classify papers by their abstract to see either they are about social media or not. To do so, the raw database of 2230 abstracts of different papers that are attached here need to be preprocessed so there will be a matrix of features to train and test our methods. In this project, we can see how to preprocess such databases and how to train a model that can predict if an unseen data -abstract of a paper- is about social media or not.